moved components to sub directory, added placeholder for group management
