+++ admin/src/components/groups.tsx
@@ -1,0 +1,23 @@
+import * as React from "react";
+import * as ReactDOM from "react-dom";
+
+import Fragment from "./fragment";
+
+const $window = window as any;
+
+/** Translates text */
+const _ = $window._ as (text: string) => string;
+
+export class Groups extends React.Component<any, any> {
+
+	constructor(props) {
+		super(props);
+	}
+
+	public render() {
+		return (
+			<div>Placeholder for group management functions</div>
+		);
+	}
+
+}
