+++ admin/src/index.test.tsx
@@ -5,7 +5,7 @@ import * as ReactDOM from "react-dom";
 // Setup Enzyme adapter
 import {configure, shallow} from "enzyme";
 import * as Adapter from "enzyme-adapter-react-16";
-configure({ adapter: new Adapter() });
+configure({ adapter: new (Adapter as any)() });
 
 // tslint:disable:no-unused-expression
 import { assert, expect } from "chai";
